restructure macro logic
