+++ crates/macros/src/wasm_export/type_utils.rs
@@ -1,0 +1,24 @@
+use syn::{Path, PathSegment, Type, TypePath};
+
+/// Try to extract the inner type from a Result<T, E> type, returning None if not a Result
+pub fn try_extract_result_inner_type(return_type: &Box<Type>) -> Option<&Type> {
+    if let Type::Path(TypePath {
+        path: Path { segments, .. },
+        ..
+    }) = &**return_type
+    {
+        if let Some(PathSegment {
+            ident, arguments, ..
+        }) = segments.first()
+        {
+            if ident.to_string() == "Result" {
+                if let syn::PathArguments::AngleBracketed(args) = arguments {
+                    if let Some(syn::GenericArgument::Type(t)) = args.args.first() {
+                        return Some(t);
+                    }
+                }
+            }
+        }
+    }
+    None
+}
